refactor: extract loadModelFile function and move utils
