+++ islamic_image_loader/src/main/java/com/baghdad/islamic_image_loader/component/SafeImage.kt
@@ -23,7 +23,7 @@ import coil3.compose.asPainter
 import coil3.toBitmap
 import com.baghdad.islamic_image_loader.R
 import com.baghdad.islamic_image_loader.model.detectFaces
-import com.baghdad.islamic_image_loader.model.imageUrlLoader
+import com.baghdad.islamic_image_loader.imageLoader.imageUrlLoader
 import com.baghdad.islamic_image_loader.model.predictGender
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
